Move Action, Duplication action under beets.importer.actions
