+++ beets/importer/tasks.py
@@ -21,12 +21,10 @@ import shutil
 import time
 from collections import defaultdict
 from collections.abc import Callable
-from enum import Enum
 from tempfile import mkdtemp
 from typing import TYPE_CHECKING, Any
 
 import mediafile
-from typing_extensions import Self
 
 from beets import config, library, plugins, util
 from beets.autotag import AlbumMatch, tag_album, tag_item
@@ -34,6 +32,7 @@ from beets.dbcore.query import PathQuery
 from beets.util import extension
 from beets.util.extension import remux_mpeglayer3_wav
 
+from .actions import Action
 from .state import ImportState
 
 if TYPE_CHECKING:
@@ -73,50 +72,8 @@ REIMPORT_FRESH_FIELDS_ALBUM = [*REIMPORT_FRESH_FIELDS_ITEM, "media"]
 log = logging.getLogger("beets")
 
 
-class DuplicateAction(str, Enum):
-    text: str
-
-    def __new__(cls, code: str, text: str) -> Self:
-        obj = str.__new__(cls, code)
-        obj._value_ = code
-        obj.text = text
-        return obj
-
-    @classmethod
-    def options(cls) -> list[str]:
-        return [d.text for d in cls]
-
-    @classmethod
-    def strict_options(cls) -> list[str]:
-        return [d.text for d in set(cls) - {DuplicateAction.ASK}]
-
-    @classmethod
-    def choices(cls) -> dict[str, str]:
-        return {d.name.lower(): d.value for d in cls}
-
-    SKIP = "s", "Skip new"
-    MERGE = "m", "Merge all"
-    REMOVE = "r", "Remove old"
-    KEEP = "k", "Keep all"
-    ASK = "a", "Ask"
-
-
 class ImportAbortError(Exception):
     """Raised when the user aborts the tagging operation."""
-
-
-class Action(Enum):
-    """Enumeration of possible actions for an import task."""
-
-    SKIP = "SKIP"
-    ASIS = "ASIS"
-    TRACKS = "TRACKS"
-    APPLY = "APPLY"
-    ALBUMS = "ALBUMS"
-    RETAG = "RETAG"
-    # The RETAG action represents "don't apply any match, but do record
-    # new metadata". It's not reachable via the standard command prompt but
-    # can be used by plugins.
 
 
 class BaseImportTask:
